Refactor: LineStations 일급 컬렉션 refactor
