+++ src/test/java/nextstep/subway/line/application/LineStationServiceTest.java
@@ -57,7 +57,7 @@ public class LineStationServiceTest {
         lineStationService.addSection(line.getId(), new SectionRequest(newStation.getId(), station1.getId(), 10));
 
         // then - 역이 3개, 역 정보 조회는 일단 대기
-        assertThat(line.getLineStations()).hasSize(3);
+        assertThat(line.getLineStations().getLineStations()).hasSize(3);
     }
 
     @Test
@@ -70,7 +70,7 @@ public class LineStationServiceTest {
         lineStationService.addSection(line.getId(), new SectionRequest(station2.getId(), newStation.getId(), 10));
 
         // then - 역이 3개, 역 정보 조회는 일단 대기
-        assertThat(line.getLineStations()).hasSize(3);
+        assertThat(line.getLineStations().getLineStations()).hasSize(3);
     }
 
     @Test
@@ -80,10 +80,10 @@ public class LineStationServiceTest {
 
         // when - 새로운 역 구간 추가 ( 하행종점 출발, 새로운 역 도착)
         Station newStation = stationRepository.save(new Station("사당역"));
-        lineStationService.addSection(line.getId(), new SectionRequest(station1.getId(), newStation.getId(), 10));
+        lineStationService.addSection(line.getId(), new SectionRequest(station1.getId(), newStation.getId(), 5));
 
         // then - 역이 3개, 역 정보 조회는 일단 대기
-        assertThat(line.getLineStations()).hasSize(3);
+        assertThat(line.getLineStations().getLineStations()).hasSize(3);
     }
 
     @Test
@@ -93,10 +93,10 @@ public class LineStationServiceTest {
 
         // when - 새로운 역 구간 추가 ( 하행종점 출발, 새로운 역 도착)
         Station newStation = stationRepository.save(new Station("사당역"));
-        lineStationService.addSection(line.getId(), new SectionRequest(newStation.getId(), station2.getId(), 10));
+        lineStationService.addSection(line.getId(), new SectionRequest(newStation.getId(), station2.getId(), 5));
 
         // then - 역이 3개, 역 정보 조회는 일단 대기
-        assertThat(line.getLineStations()).hasSize(3);
+        assertThat(line.getLineStations().getLineStations()).hasSize(3);
     }
 
     @Test
